pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ptc-agent"
version = "0.1.0"
description = "ptc-agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx[http2]>=0.28.1",
    # LangChain core packages (v1)
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-anthropic>=1.0.0",
    "langchain_google_genai>=2.0.0",
    "langchain-deepseek>=0.1.4",
    # LangGraph
    "langgraph>=0.3.5",
    "langgraph-checkpoint-postgres>=2.0.0",
    # Other dependencies
    "python-dotenv>=1.0.1",
    "crawl4ai>=0.4.0",
    "playwright",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "websockets>=12.0",
    "pandas>=2.2.3",
    "numpy>=1.24,<3.0",
    "json-repair>=0.7.0",
    "jinja2>=3.1.3",
    "mcp>=1.6.0",
    "psycopg[binary,pool]>=3.2.9",
    "alembic>=1.15.0",
    "sqlalchemy>=2.0.0",
    "yfinance==0.2.61",
    "matplotlib",
    "mplfinance==0.12.10b0",
    "tabulate",
    "scikit-learn",
    "scipy",
    "statsmodels",
    "seaborn",
    "plotly",
    "alibabacloud_oss_v2",
    "aiofiles==24.1.0",
    "setuptools",
    "rich>=14.2.0",
    "redis>=5.0.0",
    "langchain-qwq>=0.3.0",
    "edgartools>=5.7.2",
    "langsmith-fetch>=0.3.1",
    "structlog",
    "deepagents>=0.3.5",
    "daytona-sdk>=0.130.0",
    "boto3>=1.42.28",
    "tavily-python>=0.7.18",
    "PyJWT[crypto]>=2.8",
    "ptc-cli",
    "certifi>=2026.1.4",
    "requests>=2.32.5",
    "croniter>=2.0.0",
    "aiodocker>=0.23.0",
]

[tool.uv.sources]
ptc-cli = { path = "libs/ptc-cli", editable = true }

[project.optional-dependencies]
dev = [
    "ruff",
    "black>=24.2.0",
    "langgraph-cli[inmem]>=0.2.10",
]
test = [
    "pytest>=9.0.1",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=1.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ptc_agent"]

[tool.uv]
required-version = ">=0.6.15"

[tool.ruff.lint]
ignore = ["E741"]  # allow short names like l, O, I (common in OHLCV code)

[tool.ruff.lint.per-file-ignores]
"src/server/app/setup.py" = ["E402"]  # router imports intentionally placed after app creation

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "strict"
markers = [
    "integration: hits real APIs — requires env vars (deselect with '-m \"not integration\"')",
    "slow: long-running tests (stress, concurrency) -- opt-in",
    "provider_docker: tests requiring Docker daemon",
    "provider_daytona: tests requiring Daytona API key",
]
addopts = "-m 'not integration and not slow'"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
